Google api wrapping only api calling code
